+++ pkg/errorclient/README.md
@@ -1,9 +1,11 @@
+<!-- gomarkdoc:embed:start -->
+
 <!-- Code generated by gomarkdoc. DO NOT EDIT -->
 
 # errorclient
 
 ```go
-import "github.com/error-fyi/go-fyi/internal/errorclient"
+import "github.com/error-fyi/go-fyi/pkg/errorclient"
 ```
 
 ## Index
@@ -13,7 +15,7 @@ import "github.com/error-fyi/go-fyi/internal/errorclient"
 
 
 <a name="ErrClient"></a>
-## type [ErrClient](<https://github.com/error-fyi/go-fyi/blob/main/internal/errorclient/client.go#L8-L10>)
+## type [ErrClient](<https://github.com/error-fyi/go-fyi/blob/main/pkg/errorclient/client.go#L8-L10>)
 
 
 
@@ -24,7 +26,7 @@ type ErrClient interface {
 ```
 
 <a name="ErrClientOptions"></a>
-## type [ErrClientOptions](<https://github.com/error-fyi/go-fyi/blob/main/internal/errorclient/client.go#L13-L31>)
+## type [ErrClientOptions](<https://github.com/error-fyi/go-fyi/blob/main/pkg/errorclient/client.go#L13-L31>)
 
 ErrClientOptions for the error handler. Use it to configure the aloe error handler
 
@@ -51,3 +53,6 @@ type ErrClientOptions struct {
 ```
 
 Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
+
+
+<!-- gomarkdoc:embed:end -->
