+++ pkg/errorclient/local/README.md
@@ -1,9 +1,11 @@
+<!-- gomarkdoc:embed:start -->
+
 <!-- Code generated by gomarkdoc. DO NOT EDIT -->
 
 # local
 
 ```go
-import "github.com/error-fyi/go-fyi/internal/errorclient/local"
+import "github.com/error-fyi/go-fyi/pkg/errorclient/local"
 ```
 
 ## Index
@@ -25,7 +27,7 @@ var (
 ```
 
 <a name="New"></a>
-## func [New](<https://github.com/error-fyi/go-fyi/blob/main/internal/errorclient/local/local.go#L31>)
+## func [New](<https://github.com/error-fyi/go-fyi/blob/main/pkg/errorclient/local/local.go#L31>)
 
 ```go
 func New(opts errorclient.ErrClientOptions) errorclient.ErrClient
@@ -34,7 +36,7 @@ func New(opts errorclient.ErrClientOptions) errorclient.ErrClient
 
 
 <a name="Client"></a>
-## type [Client](<https://github.com/error-fyi/go-fyi/blob/main/internal/errorclient/local/local.go#L17-L20>)
+## type [Client](<https://github.com/error-fyi/go-fyi/blob/main/pkg/errorclient/local/local.go#L17-L20>)
 
 
 
@@ -46,7 +48,7 @@ type Client struct {
 ```
 
 <a name="Client.GenerateErrorMessageFromCode"></a>
-### func \(\*Client\) [GenerateErrorMessageFromCode](<https://github.com/error-fyi/go-fyi/blob/main/internal/errorclient/local/local.go#L47>)
+### func \(\*Client\) [GenerateErrorMessageFromCode](<https://github.com/error-fyi/go-fyi/blob/main/pkg/errorclient/local/local.go#L47>)
 
 ```go
 func (l *Client) GenerateErrorMessageFromCode(ctx context.Context, code string) (string, error)
@@ -55,3 +57,6 @@ func (l *Client) GenerateErrorMessageFromCode(ctx context.Context, code string) 
 
 
 Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
+
+
+<!-- gomarkdoc:embed:end -->
