+++ pkg/types/README.md
@@ -1,9 +1,11 @@
+<!-- gomarkdoc:embed:start -->
+
 <!-- Code generated by gomarkdoc. DO NOT EDIT -->
 
 # types
 
 ```go
-import "github.com/error-fyi/go-fyi/internal/types"
+import "github.com/error-fyi/go-fyi/pkg/types"
 ```
 
 ## Index
@@ -11,3 +13,6 @@ import "github.com/error-fyi/go-fyi/internal/types"
 
 
 Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
+
+
+<!-- gomarkdoc:embed:end -->
