+++ pkg/types/v0.1.0/README.md
@@ -1,9 +1,11 @@
+<!-- gomarkdoc:embed:start -->
+
 <!-- Code generated by gomarkdoc. DO NOT EDIT -->
 
 # v0\_1\_0
 
 ```go
-import "github.com/error-fyi/go-fyi/internal/types/v0.1.0"
+import "github.com/error-fyi/go-fyi/pkg/types/v0.1.0"
 ```
 
 ## Index
@@ -24,7 +26,7 @@ import "github.com/error-fyi/go-fyi/internal/types/v0.1.0"
 
 
 <a name="Error"></a>
-## type [Error](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L9-L30>)
+## type [Error](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L9-L30>)
 
 
 
@@ -54,7 +56,7 @@ type Error struct {
 ```
 
 <a name="Error.UnmarshalJSON"></a>
-### func \(\*Error\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L121>)
+### func \(\*Error\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L121>)
 
 ```go
 func (j *Error) UnmarshalJSON(b []byte) error
@@ -63,7 +65,7 @@ func (j *Error) UnmarshalJSON(b []byte) error
 UnmarshalJSON implements json.Unmarshaler.
 
 <a name="ErrorDefinitions"></a>
-## type [ErrorDefinitions](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L32>)
+## type [ErrorDefinitions](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L32>)
 
 
 
@@ -72,7 +74,7 @@ type ErrorDefinitions map[string]Error
 ```
 
 <a name="ErrorMeta"></a>
-## type [ErrorMeta](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L35-L38>)
+## type [ErrorMeta](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L35-L38>)
 
 Metadata information about the error.
 
@@ -84,7 +86,7 @@ type ErrorMeta struct {
 ```
 
 <a name="ErrorMetaLoc"></a>
-## type [ErrorMetaLoc](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L40-L46>)
+## type [ErrorMetaLoc](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L40-L46>)
 
 
 
@@ -99,7 +101,7 @@ type ErrorMetaLoc struct {
 ```
 
 <a name="ErrorMetaLoc.UnmarshalJSON"></a>
-### func \(\*ErrorMetaLoc\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L154>)
+### func \(\*ErrorMetaLoc\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L154>)
 
 ```go
 func (j *ErrorMetaLoc) UnmarshalJSON(b []byte) error
@@ -108,7 +110,7 @@ func (j *ErrorMetaLoc) UnmarshalJSON(b []byte) error
 UnmarshalJSON implements json.Unmarshaler.
 
 <a name="ErrorSeverity"></a>
-## type [ErrorSeverity](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L48>)
+## type [ErrorSeverity](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L48>)
 
 
 
@@ -135,7 +137,7 @@ const ErrorSeveritySevere ErrorSeverity = "severe"
 ```
 
 <a name="ErrorSeverity.UnmarshalJSON"></a>
-### func \(\*ErrorSeverity\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L55>)
+### func \(\*ErrorSeverity\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L55>)
 
 ```go
 func (j *ErrorSeverity) UnmarshalJSON(b []byte) error
@@ -144,7 +146,7 @@ func (j *ErrorSeverity) UnmarshalJSON(b []byte) error
 UnmarshalJSON implements json.Unmarshaler.
 
 <a name="Manifest"></a>
-## type [Manifest](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L174-L195>)
+## type [Manifest](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L174-L195>)
 
 
 
@@ -174,7 +176,7 @@ type Manifest struct {
 ```
 
 <a name="Manifest.UnmarshalJSON"></a>
-### func \(\*Manifest\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L198>)
+### func \(\*Manifest\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L198>)
 
 ```go
 func (j *Manifest) UnmarshalJSON(b []byte) error
@@ -183,7 +185,7 @@ func (j *Manifest) UnmarshalJSON(b []byte) error
 UnmarshalJSON implements json.Unmarshaler.
 
 <a name="Suggestion"></a>
-## type [Suggestion](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L74-L86>)
+## type [Suggestion](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L74-L86>)
 
 
 
@@ -204,7 +206,7 @@ type Suggestion struct {
 ```
 
 <a name="Suggestion.UnmarshalJSON"></a>
-### func \(\*Suggestion\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L89>)
+### func \(\*Suggestion\) [UnmarshalJSON](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L89>)
 
 ```go
 func (j *Suggestion) UnmarshalJSON(b []byte) error
@@ -213,7 +215,7 @@ func (j *Suggestion) UnmarshalJSON(b []byte) error
 UnmarshalJSON implements json.Unmarshaler.
 
 <a name="Suggestions"></a>
-## type [Suggestions](<https://github.com/error-fyi/go-fyi/blob/main/internal/types/v0.1.0/types.go#L112>)
+## type [Suggestions](<https://github.com/error-fyi/go-fyi/blob/main/pkg/types/v0.1.0/types.go#L112>)
 
 
 
@@ -222,3 +224,6 @@ type Suggestions map[string]Suggestion
 ```
 
 Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
+
+
+<!-- gomarkdoc:embed:end -->
